add new camera node
