+++ nodes/camera-legacy/icons/camera-take-picture.png
