+++ nodes/camera-legacy/icons/preview.png
