+++ nodes/camera-legacy/icons/switch.png
