+++ nodes/camera-legacy/locales/en-US/camera.html
@@ -1,0 +1,24 @@
+<script type="text/x-red" data-help-name="camera-open">
+    <p>Starts the camera.</p>
+    <p><b>Note</b>: It cannot be executed if the app (redmobile) is in the background.</p>
+</script>
+
+<script type="text/x-red" data-help-name="camera-close">
+    <p>Stop the camera.</p>
+    <p><b>Note</b>: It cannot be executed if the app (redmobile) is in the background.</p>
+</script>
+
+<script type="text/x-red" data-help-name="take-picture">
+    <p>Take the picture.</p>
+    <h3>Outputs</h3>
+    <dl class="message-properties">
+        <dt>payload<span class="property-type">string</span></dt>
+        <dd>base64 encoded jpeg image</dd>
+    </dl>
+    <p><b>Note</b>: It cannot be executed if the app (redmobile) is in the background.</p>
+</script>
+
+<script type="text/x-red" data-help-name="camera-switch">
+    <p>Switch between the rear camera and front camera, if available.</p>
+    <p><b>Note</b>: It cannot be executed if the app (redmobile) is in the background.</p>
+</script>
