+++ nodes/camera-legacy/locales/en-US/camera.json
@@ -1,0 +1,38 @@
+{
+    "camera-open":{
+        "camera-open": "Camera Open",
+        "label":{
+            "preview": "Preview",
+            "direction": "Direction"
+        },
+        "preview":{
+            "enable": "Enable",
+            "disable": "Disable"
+        },
+        "direction":{
+            "front": "Front",
+            "back": "Back"
+        },
+        "errors":{
+            "response": "Could not open camera."
+        }
+    },
+    "take-picture":{
+        "take-picture": "Take Picture",
+        "errors":{
+            "response": "Could not take Picture."
+        }
+    },
+    "camera-close":{
+        "camera-close": "Camera Close",
+        "errors":{
+            "response": "Could not close camera."
+        }
+    },
+    "camera-switch":{
+        "camera-switch": "Camera Switch",
+        "errors":{
+            "response": "Could not switch camera."
+        }
+    }
+}
