+++ nodes/qrcode/locales/en-US/scan.html
@@ -1,4 +1,3 @@
-
 <script type="text/x-red" data-help-name="qrcode scan">
     <p>Scan a QR code (barcode) and get the result.</p>
     <h3>Outputs</h3>
